# api/openapi.yaml
openapi: 3.0.1
info:
  title: eMekdep API
  description: ''
  version: 1.0.0
...
paths:
...
  /contact-items:
    get:
      summary: /contact-items
      deprecated: false
      description: ''
      tags:
        - School
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      contact_items:
                        type: array
                        items:
                          $ref: '#/components/schemas/ContactItemList'
                    type: object
      security: []
    post:
      summary: /contact-items
      deprecated: false
      description: ''
      tags:
        - School
      parameters: []
      requestBody:
        content:
          application/json:
            schema: &ref_19
              $ref: '#/components/schemas/ContactItemCreate'
            example: ''
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      data:
                        type: object
                        properties:
                          item: &ref_18
                            $ref: '#/components/schemas/ContactItem'
                    type: object
      security: []
    delete:
      summary: /contact-items
      deprecated: false
      description: ''
      tags:
        - School
      parameters:
        - name: ids
          in: query
          description: ''
          required: false
          schema:
            type: array
            items:
              type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      data:
                        type: object
                        properties:
                          classrooms:
                            type: array
                            items: *ref_18
                    type: object
      security: []
...
  /reports/data:
    get:
      summary: /reports/data
      deprecated: false
      description: ''
      tags:
        - Reports
      parameters:
        - name: start_date
          in: query
          description: ''
          required: true
          schema:
            type: string
        - name: end_date
          in: query
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      report: &ref_32
                        $ref: '#/components/schemas/ReportData'
                    type: object
      security: []
  /reports/attendance:
    get:
      summary: /reports/attendance
      deprecated: false
      description: ''
      tags:
        - Reports
      parameters:
        - name: start_date
          in: query
          description: ''
          required: true
          schema:
            type: string
        - name: end_date
          in: query
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      report: *ref_32
                    type: object
      security: []
...
  /reports/period_finished:
    get:
      summary: /reports/period_finished
      deprecated: false
      description: ''
      tags:
        - Reports
      parameters:
        - name: period_number
          in: query
          description: ''
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      report: *ref_32
                    type: object
      security: []
...
components:
  schemas:
...
    ContactItemCreate:
      properties:
        id: *ref_0
        related_ids:
          type: array
          items: *ref_0
        message:
          type: string
        type: &ref_48
          $ref: '#/components/schemas/ContactItemType'
        status: &ref_49
          $ref: '#/components/schemas/ContactItemStatus'
        files:
          type: array
          items: *ref_36
      type: object
...
    ReportRow:
      properties:
        region:
          type: string
        region_code:
          type: string
        school:
          type: string
        schools_code:
          type: string
        school_id:
          type: integer
        classroom:
          type: string
        classroom_id:
          type: integer
        user:
          type: string
        user_id:
          type: integer
        percent:
          type: integer
        values:
          type: array
          items:
            type: string
      type: object
    ReportTotal:
      properties:
        title:
          type: string
        value:
          type: string
        type:
          type: string
      type: object
    ReportData:
      properties:
        has_between_date:
          type: boolean
          examples:
            - false
        has_detail:
          type: boolean
          examples:
            - false
        has_period_number:
          type: boolean
          examples:
            - false
        headers:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            $ref: '#/components/schemas/ReportRow'
        totals:
          type: array
          items:
            $ref: '#/components/schemas/ReportTotal'
      type: object
...
    ContactItemList:
      allOf:
        - *ref_18
        - properties:
            user:
              type: object
              default: null
              properties: {}
              nullable: true
            school:
              type: object
              default: null
              properties: {}
              nullable: true
          type: object
    ContactItem:
      properties:
        id: *ref_0
        related_ids:
          type: array
          items: *ref_0
        message:
          type: string
        type: *ref_48
        status: *ref_49
        files:
          type: array
          items: *ref_36
        user: *ref_9
        school: *ref_50
      type: object
...
    ContactItemStatus:
      type: string
      enum:
        - waiting
        - todo
        - processing
        - done
        - backlog
        - rejected
    ContactItemType:
      type: string
      enum:
        - review
        - complaint
        - suggestion
        - data_complaint
...
  securitySchemes: {}
servers: []